create separate Account comp
