+++ src/components/commonComponents/Account/Account.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import "./Account.css";
+
+export default function Account() {
+  return <div>Account</div>;
+}
