+++ src/components/commonComponents/SignInSignUpButton/SignInSignUpButton.jsx
@@ -1,13 +1,14 @@
 import React from "react";
-import "./LoginButtonForHeader.css";
+import "./SignInSignUpButton.css";
 import { motion } from "framer-motion";
 
-export default function LoginButtonForHeader({ onClickHandler, text }) {
+export default function SignInSignUpButton({ onClickHandler, text }) {
   return (
     <div>
       <motion.button
         className="signIn-signUp-button-header"
         type="button"
+        initial={{ textShadow: "0px 0px 15px rgb(255, 255, 255)" }}
         whileHover={{
           scale: 1.1,
           textShadow: "0px 0px 8px rgb(255, 255, 255)",
